move quantization and lora to new home
