+++ tests/unit_tests/cpu/test_quantization.py
@@ -17,14 +17,6 @@ from torchtitan.components.data import (
     SingleDatasetConfig,
 )
 from torchtitan.components.data.sources import HuggingFaceRandomAccessSource
-from torchtitan.components.quantization import Float8Linear, Float8LinearConverter
-from torchtitan.components.quantization.float8 import _get_float8_grouped_experts_cls
-from torchtitan.components.quantization.mxfp8.converter import (
-    _get_mxfp8_grouped_experts_cls,
-    MXFP8Linear,
-    MXFP8LinearConverter,
-)
-from torchtitan.components.quantization.utils import has_quantization
 from torchtitan.config import ConfigManager
 from torchtitan.models.common.config_utils import make_router_config
 from torchtitan.models.common.decoder_sharding import colwise_config, rowwise_config
@@ -32,6 +24,14 @@ from torchtitan.models.common.feed_forward import FeedForward
 from torchtitan.models.common.linear import Linear
 from torchtitan.models.common.moe import GroupedExperts
 from torchtitan.models.gpt_oss.moe import GptOssGroupedExperts
+from torchtitan.quantization import Float8Linear, Float8LinearConverter
+from torchtitan.quantization.float8 import _get_float8_grouped_experts_cls
+from torchtitan.quantization.mxfp8.converter import (
+    _get_mxfp8_grouped_experts_cls,
+    MXFP8Linear,
+    MXFP8LinearConverter,
+)
+from torchtitan.quantization.utils import has_quantization
 
 
 def test_no_float8_by_default():
@@ -70,7 +70,7 @@ def test_mxfp8_converter_rejects_router_gate(monkeypatch):
     pytest.importorskip("torchao")
     if MXFP8Linear is None:
         pytest.skip("torchao MXFP8Linear is unavailable")
-    import torchtitan.components.quantization.mxfp8.converter as converter_mod
+    import torchtitan.quantization.mxfp8.converter as converter_mod
 
     monkeypatch.setattr(converter_mod, "has_cuda_capability", lambda *_: True)
     converter = MXFP8LinearConverter(MXFP8LinearConverter.Config())
@@ -80,11 +80,11 @@ def test_mxfp8_converter_rejects_router_gate(monkeypatch):
 
 def test_nvfp4_converter_rejects_router_gate(monkeypatch):
     pytest.importorskip("torchao")
-    from torchtitan.components.quantization import NVFP4Linear, NVFP4LinearConverter
+    from torchtitan.quantization import NVFP4Linear, NVFP4LinearConverter
 
     if NVFP4Linear is None:
         pytest.skip("torchao NVFP4 training prototype not available")
-    import torchtitan.components.quantization.nvfp4 as nvfp4_mod
+    import torchtitan.quantization.nvfp4 as nvfp4_mod
 
     monkeypatch.setattr(nvfp4_mod, "has_cuda_capability", lambda *_: True)
     converter = NVFP4LinearConverter(NVFP4LinearConverter.Config())
@@ -130,14 +130,14 @@ def test_nvfp4_converter_targets_layers_not_lm_head(
     monkeypatch, module, recipe, expected_num_layers
 ):
     pytest.importorskip("torchao")
-    from torchtitan.components.quantization import NVFP4Linear
+    from torchtitan.quantization import NVFP4Linear
 
     if NVFP4Linear is None:
         pytest.skip("torchao NVFP4 training prototype not available")
     # Exercise convert() targeting independent of GPU: bypass the sm100 gate
     # that NVFP4LinearConverter.__init__ enforces (hardware is irrelevant to the
     # config-tree transform under test).
-    import torchtitan.components.quantization.nvfp4 as nvfp4_mod
+    import torchtitan.quantization.nvfp4 as nvfp4_mod
 
     monkeypatch.setattr(nvfp4_mod, "has_cuda_capability", lambda *_: True)
 
@@ -160,7 +160,7 @@ def test_nvfp4_converter_targets_layers_not_lm_head(
 
 
 def test_nvfp4_bf16_tail_fqns():
-    from torchtitan.components.quantization.nvfp4 import nvfp4_bf16_tail_fqns
+    from torchtitan.quantization.nvfp4 import nvfp4_bf16_tail_fqns
 
     # 32 layers, 15% tail -> ceil(4.8)=5 bf16, convert layers 0..26.
     fqns = nvfp4_bf16_tail_fqns(32, 0.15)
@@ -194,13 +194,13 @@ def test_nvfp4_first_85_pct_layers_converts_only_leading_layers(
     monkeypatch, module, recipe, expected_cutoff
 ):
     pytest.importorskip("torchao")
-    from torchtitan.components.quantization import NVFP4Linear
+    from torchtitan.quantization import NVFP4Linear
 
     if NVFP4Linear is None:
         pytest.skip("torchao NVFP4 training prototype not available")
     import math
 
-    import torchtitan.components.quantization.nvfp4 as nvfp4_mod
+    import torchtitan.quantization.nvfp4 as nvfp4_mod
 
     monkeypatch.setattr(nvfp4_mod, "has_cuda_capability", lambda *_: True)
 
@@ -229,7 +229,7 @@ def test_nvfp4_first_85_pct_layers_converts_only_leading_layers(
 
 def _nvfp4_linear_cls():
     pytest.importorskip("torchao")
-    from torchtitan.components.quantization import NVFP4Linear
+    from torchtitan.quantization import NVFP4Linear
 
     if NVFP4Linear is None:
         pytest.skip("torchao NVFP4 training prototype not available")
@@ -296,7 +296,7 @@ def test_nvfp4_build_configures_local_spmd_sharding(sharding_config_factory, inp
 )
 def test_nvfp4_recipes_parse(monkeypatch, module, recipe):
     _nvfp4_linear_cls()
-    import torchtitan.components.quantization.nvfp4 as nvfp4_mod
+    import torchtitan.quantization.nvfp4 as nvfp4_mod
 
     monkeypatch.setattr(nvfp4_mod, "has_cuda_capability", lambda *_: True)
     base_args = ["--module", module, "--config", recipe]
@@ -314,7 +314,7 @@ def test_nvfp4_recipes_parse(monkeypatch, module, recipe):
 )
 def test_qwen3_recipes_resolve(monkeypatch, recipe):
     _nvfp4_linear_cls()
-    import torchtitan.components.quantization.nvfp4 as nvfp4_mod
+    import torchtitan.quantization.nvfp4 as nvfp4_mod
 
     monkeypatch.setattr(nvfp4_mod, "has_cuda_capability", lambda *_: True)
     config = ConfigManager().parse_args(["--module", "qwen3", "--config", recipe])
@@ -338,7 +338,7 @@ def test_nvfp4_module_buffers_and_native_checkpoint():
     constant and the SR seed is per-rank -- so a native checkpoint carries only
     the stock weight."""
     NVFP4Linear = _nvfp4_linear_cls()
-    from torchtitan.components.quantization.nvfp4 import _HARDCODED_SIGN_VECTOR
+    from torchtitan.quantization.nvfp4 import _HARDCODED_SIGN_VECTOR
 
     module = NVFP4Linear.Config(in_features=512, out_features=1024).build()
     assert {name for name, _ in module.named_parameters()} == {"weight"}
@@ -373,7 +373,7 @@ def test_nvfp4_stock_checkpoint_loads_before_init_states():
 def test_nvfp4_hf_export_strips_buffers(monkeypatch):
     """The HF export boundary contains only stock keys -- no NVFP4 runtime buffers."""
     NVFP4Linear = _nvfp4_linear_cls()
-    import torchtitan.components.quantization.nvfp4 as nvfp4_mod
+    import torchtitan.quantization.nvfp4 as nvfp4_mod
 
     monkeypatch.setattr(nvfp4_mod, "has_cuda_capability", lambda *_: True)
     from torchtitan.models.llama3.state_dict_adapter import Llama3StateDictAdapter
@@ -491,8 +491,8 @@ def test_mxfp8_linear_validates_config_and_installs_weight_wrapper():
     pytest.importorskip("torchao")
     if MXFP8Linear is None:
         pytest.skip("torchao MXFP8Linear is unavailable")
-    from torchtitan.components.quantization._fsdp_tensor import _UnshardedFSDPTensor
-    from torchtitan.components.quantization.mxfp8.tensor import (
+    from torchtitan.quantization._fsdp_tensor import _UnshardedFSDPTensor
+    from torchtitan.quantization.mxfp8.tensor import (
         _LinearShardedTensorWithMXFP8Compute,
     )
 
@@ -533,7 +533,7 @@ def test_mxfp8_converter_replaces_a_root_linear_config(monkeypatch):
     the one branch that has to return the replacement. Not covered by the FQN
     test below, which passes a FeedForward and so always has a parent.
     """
-    import torchtitan.components.quantization.mxfp8.converter as converter_mod
+    import torchtitan.quantization.mxfp8.converter as converter_mod
 
     monkeypatch.setattr(converter_mod, "has_cuda_capability", lambda *_: True)
     converter = MXFP8LinearConverter(
@@ -551,7 +551,7 @@ def test_mxfp8_converter_replaces_a_root_linear_config(monkeypatch):
 
 
 def test_mxfp8_converter_applies_mxfp8_saved_input_fqns(monkeypatch):
-    import torchtitan.components.quantization.mxfp8.converter as converter_mod
+    import torchtitan.quantization.mxfp8.converter as converter_mod
 
     monkeypatch.setattr(converter_mod, "has_cuda_capability", lambda *_: True)
     converter = MXFP8LinearConverter(
@@ -577,7 +577,7 @@ def test_mxfp8_converter_applies_mxfp8_saved_input_fqns(monkeypatch):
 
 
 def test_mxfp8_converter_rejects_unmatched_saved_input_fqns(monkeypatch):
-    import torchtitan.components.quantization.mxfp8.converter as converter_mod
+    import torchtitan.quantization.mxfp8.converter as converter_mod
 
     monkeypatch.setattr(converter_mod, "has_cuda_capability", lambda *_: True)
     converter = MXFP8LinearConverter(
@@ -633,7 +633,7 @@ def test_builtin_mxfp8_configs_assign_input_activation_format_for_backward(
 ):
     if MXFP8Linear is None:
         pytest.skip("torchao MXFP8Linear is unavailable")
-    import torchtitan.components.quantization.mxfp8.converter as converter_mod
+    import torchtitan.quantization.mxfp8.converter as converter_mod
 
     monkeypatch.setattr(converter_mod, "has_cuda_capability", lambda *_: True)
     if config_factory == "llama3":
@@ -675,7 +675,7 @@ def test_mxfp8_linear_loads_stock_checkpoint():
     pytest.importorskip("torchao")
     if MXFP8Linear is None:
         pytest.skip("torchao MXFP8Linear is unavailable")
-    from torchtitan.components.quantization.mxfp8.tensor import (
+    from torchtitan.quantization.mxfp8.tensor import (
         _LinearShardedTensorWithMXFP8Compute,
     )
