+++ torchtitan/config/transform/__init__.py
@@ -9,6 +9,7 @@
 from .apply import apply_transforms, transform_model_config_
 from .base import convert_config_type, ModelConfigTransform
 from .context_parallel import ContextParallelTransform
+from .lora import LoRAConverter
 
 __all__ = [
     "ModelConfigTransform",
@@ -16,4 +17,5 @@ __all__ = [
     "transform_model_config_",
     "convert_config_type",
     "ContextParallelTransform",
+    "LoRAConverter",
 ]
