+++ torchtitan/models/flux/config_registry.py
@@ -10,7 +10,6 @@ from torchtitan.components.checkpointer import CheckpointManager
 from torchtitan.components.data import GrainDataLoader, SingleDatasetConfig
 from torchtitan.components.loss import MSELoss
 from torchtitan.components.optimizer import default_adamw, LRSchedulersContainer
-from torchtitan.components.quantization import MXFP8LinearConverter
 from torchtitan.config import CompileConfig, ParallelismConfig, TrainingConfig
 from torchtitan.distributed.activation_checkpoint import FullAC
 from torchtitan.models.flux.configs import FluxEncoderConfig, Inference, SamplingConfig
@@ -29,6 +28,7 @@ from torchtitan.models.flux.utils import (
 )
 from torchtitan.models.flux.validate import FluxValidator
 from torchtitan.observability.metrics import MetricsProcessor
+from torchtitan.quantization import MXFP8LinearConverter
 
 from . import model_registry
