+++ torchtitan/models/llama3/config_registry.py
@@ -16,12 +16,6 @@ from torchtitan.components.data import (
 )
 from torchtitan.components.loss import ChunkedLossWrapper, CrossEntropyLoss
 from torchtitan.components.optimizer import default_adamw, LRSchedulersContainer
-from torchtitan.components.quantization import (
-    Float8LinearConverter,
-    MXFP8LinearConverter,
-    NVFP4LinearConverter,
-)
-from torchtitan.components.quantization.nvfp4 import nvfp4_bf16_tail_fqns
 from torchtitan.components.validate import Validator
 from torchtitan.config import CompileConfig, ParallelismConfig, TrainingConfig
 from torchtitan.distributed.activation_checkpoint import FullAC, SelectiveAC
@@ -32,6 +26,12 @@ from torchtitan.models.common.config_utils import (
 )
 from torchtitan.observability.metrics import MetricsProcessor
 from torchtitan.observability.profiler import Profiler
+from torchtitan.quantization import (
+    Float8LinearConverter,
+    MXFP8LinearConverter,
+    NVFP4LinearConverter,
+)
+from torchtitan.quantization.nvfp4 import nvfp4_bf16_tail_fqns
 from torchtitan.trainer import Trainer
 
 from . import model_registry
@@ -220,7 +220,7 @@ def llama3_debugmodel_first_85_pct_layers_nvfp4(
 def llama3_debugmodel_float8_emulate_lora(
     seq_len: int | None = DEFAULT_DEBUG_MODEL_SEQ_LEN,
 ) -> Trainer.Config:
-    from torchtitan.components.lora import LoRAConverter
+    from torchtitan.config.transform import LoRAConverter
 
     config = llama3_debugmodel(seq_len=seq_len)
     config.model_spec = model_registry(
