+++ torchtitan/quantization/float8.md
@@ -13,7 +13,7 @@ Quantization is applied at config time in your `model_registry()` function via t
 
 For float8 with rowwise scaling, configure it in your config_registry function:
 ```python
-from torchtitan.components.quantization import Float8LinearConverter
+from torchtitan.quantization import Float8LinearConverter
 
 # In your model_registry call:
 model_spec = model_registry(
@@ -34,7 +34,7 @@ model_spec = model_registry(
 
 For float8 MoE expert quantization (grouped GEMMs), use `Float8GroupedExpertsConverter`:
 ```python
-from torchtitan.components.quantization import Float8LinearConverter, Float8GroupedExpertsConverter
+from torchtitan.quantization import Float8LinearConverter, Float8GroupedExpertsConverter
 
 model_spec = model_registry(
     "671B",
